Use debug-levels for debugging
Essentially this should bring some sort of timestamps for the output of the labeler, which is
extremely important when actually using the messages for debugging.

The choice for 'debug' as logging library was mostly ad-hoc, it's not completely clear which
library is the 'best' logging library.

Note that due to https://github.com/Microsoft/vscode/issues/19750 one needs to use the integrated
terminal for debugging from Visual Studio Code.